better editModes

+++ src/lib/Order.svelte
@@ -10,13 +10,13 @@
     return teams.find((team: any) => team.Id === teamId).name;
   }
 
-  let hovering = false;
-
   let isDragging = false;
   let startIndex;
   let currentIndex;
 
-  function handleTouchStart(event, index) {
+  function handleDragStart(event, index) {
+    if (!editMode) return;
+
     isDragging = true;
     startIndex = index;
     currentIndex = index;
@@ -24,7 +24,8 @@
     event.dataTransfer.setData("text/plain", index.toString());
   }
 
-  function handleTouchMove(event, index) {
+  function handleDragMove(event, index) {
+    if (!editMode) return;
     if (!isDragging) return;
 
     currentIndex = index;
@@ -32,7 +33,7 @@
     event.preventDefault();
   }
 
-  function handleTouchEnd() {
+  function handleDragEnd() {
     if (!isDragging) return;
 
     isDragging = false;
@@ -53,30 +54,32 @@
     }
   }
 
+  let deleteMode = false;
   let editMode = false;
 </script>
 
-<div><i class="bi bi-info-circle" /> Spiele per Drag and Drop anordnen.</div>
-<div>
-  <i class="bi bi-info-circle" /> Mannschaften durch Drücken auf
-  <span class="fw-bold">--</span> tauschen.
-</div>
+{#if editMode}
+  <div><i class="bi bi-info-circle" /> Spiele per Drag and Drop anordnen.</div>
+{/if}
+<button class="btn btn-primary my-1" on:click={() => (editMode = !editMode)}
+  >{editMode ? "Fertig angeordnet" : "anordnen"}</button
+>
 <button
-  class={`btn btn-outline-${editMode ? "primary" : "danger"}`}
-  on:click={() => (editMode = !editMode)}
-  >{editMode
-    ? "Löschen beenden"
-    : "einzelne Spiele entfernen (Achtung, wird die Spielfolge überschreiben!)"}</button
+  class={`btn btn-outline-${deleteMode ? "primary" : "danger"} my-1`}
+  on:click={() => (deleteMode = !deleteMode)}
+  >{deleteMode
+    ? "Entfernen beenden"
+    : "einzelne Spiele entfernen"}</button
 >
 
 <hr />
 
 <div class="d-none d-sm-block">
   <div class="row">
-    <div class="col-5" />
+    <div class={`col-${deleteMode ? "5" : "6"}`} />
     <div class="col-2 text-center">Halbzeit</div>
     <div class="col-2 text-center">Endstand</div>
-    <div class="col-2 text-center">Kommisär</div>
+    <div class="col-2 text-center">Kommissär</div>
   </div>
 </div>
 
@@ -84,32 +87,39 @@
   {#each list as match, index (`match-${index}`)}
     <div
       class="row listItem"
-      draggable="true"
+      draggable={editMode}
       class:dragging={isDragging && currentIndex === index}
-      on:touchstart={(e) => handleTouchStart(e, index)}
-      on:touchmove={(e) => handleTouchMove(e, index)}
-      on:touchend={handleTouchEnd}
-      on:dragstart={(e) => handleTouchStart(e, index)}
-      on:dragover={(e) => handleTouchMove(e, index)}
-      on:dragend={handleTouchEnd}
+      on:touchstart={(e) => handleDragStart(e, index)}
+      on:touchmove={(e) => handleDragMove(e, index)}
+      on:touchend={handleDragEnd}
+      on:dragstart={(e) => handleDragStart(e, index)}
+      on:dragover={(e) => handleDragMove(e, index)}
+      on:dragend={handleDragEnd}
     >
       <div class="col-1">{index + 1}.</div>
 
-      <div class="col-2">{getTeamNameById(match.Team1Id)}</div>
       <div class="col-2">
-        <span
-          id="switch"
-          on:click={() => {
-            match.switchTeams();
-            // needed as manual state reload
-            match.Referee = match.Referee;
-          }}
-          on:mouseenter={() => (hovering = true)}
-          on:mouseleave={() => (hovering = false)}
-          on:keydown={() => console.log("Das ist kein Button")}
-          >{hovering ? "<" : ""}--{hovering ? ">" : ""}</span
-        >
-        {getTeamNameById(match.Team2Id)}
+        {getTeamNameById(match.Team1Id) === ""
+          ? "[Bitte Teamname eingeben]"
+          : getTeamNameById(match.Team1Id)}
+      </div>
+      <div class={`col-${deleteMode ? "2" : "3"}`}>
+        {#if editMode}
+          <i
+            class="bi bi-arrow-left-right"
+            on:click={() => {
+              match.switchTeams();
+              // needed as manual state reload
+              match.Referee = match.Referee;
+            }}
+            on:keydown={() => console.log("Das ist kein Button")}
+          />
+        {:else}
+          <span>-</span>
+        {/if}
+        {getTeamNameById(match.Team2Id) === ""
+          ? "[Bitte Teamname eingeben]"
+          : getTeamNameById(match.Team2Id)}
       </div>
 
       <div class="col-2">
@@ -154,18 +164,20 @@
         <input
           type="text"
           class="form-control"
-          placeholder="Kommisär"
+          placeholder="Kommissär"
           bind:value={match.Referee}
         />
       </div>
 
-      <div class="col-1">
-        <button
-          class="btn btn-danger"
-          on:click={() => dispatch("removeMatch", { matchId: match.Id })}
-          disabled={!editMode}>X</button
-        >
-      </div>
+      {#if deleteMode}
+        <div class="col-1">
+          <button
+            class="btn btn-danger"
+            on:click={() => dispatch("removeMatch", { matchId: match.Id })}
+            >X</button
+          >
+        </div>
+      {/if}
 
       <hr />
     </div>
@@ -184,12 +196,7 @@
   }
 
   .listItem.dragging {
-    /* opacity: 0.5; */
+    opacity: 0.8;
     background-color: #0b5dd5;
-  }
-
-  #switch {
-    cursor: pointer;
-    font-weight: bold;
   }
 </style>
